seperated the card class into its own file
